modified input type to double

+++ square_with_for_statement.c
@@ -4,14 +4,14 @@
 
 int main(void)
 {
-    int i, n;
+    double i, n;
 
     printf("This program prints a table of squares.\n");
     printf("Enter number of entries in table: ");
-    scanf("%d", &n);
+    scanf("%lf", &n);
 
     for (i = 1; i <= n; ++i) {
-        printf("%10d%10d\n", i, i * i);
+        printf("%10ld%10ld\n", i, i * i);
     }
 
     return 0;
